rename pdSeries clear name decorator and move it to pvlib.tools
